Rename nbscuid -> cupid
Also renamed the conda environments, and moved more directory locations from
scratch to campaign
