+++ cupid/dev-environment.yml
@@ -1,4 +1,4 @@
-name: nbscuid-dev
+name: cupid-dev
 dependencies:
     - python==3.11.4
     - pip
