+++ cupid/quickstart.py
