+++ cupid/read.py
