+++ cupid/util.py
